added faster solution to the listOpt problem
problem: parsing a list of optional statements without splitting to much

the previous solution would first split and after parsing both alternatives throw away one alternative. this means that some code fragments can be parsed quite often which slows down the process

the new solution applies an optimistic strategy. it first parses the upcoming
statement without splitting, but already with the context of the next token.
if this yields a single result before any conflicting token, hurray, otherwise we need to use a fallback strategy (either the old strategy or the real fallback strategy of not merging before the end of the list)