Preprocessor: fix #error and note a bug
When the preprocessor copied #error to the output, it forgot a newline; a #endif
statement on the next line would thus become part of the error message!

Moreover, #if and so on are not handled correctly when parsing macro arguments;
in particular, if 'if' is a macro (as can happen in the Linux kernel), it is
expanded within #if!