FeatureExprLib - Seq representation: abstract used type, switch to ArrayBuffer of specific size
A SmallList object creates Lists needed by FeatureExprLib. Vector is immutable,
but wastes too much space and is not configurable (internal arrays are always of
size 32); therefore, use an ArrayBuffer, even if it is mutable (I checked some
immutable operations, and they seem to correctly avoid modifying the current
data structure).
Additionally, initially size the ArrayBuffer to the number of used items, rather
than the default (16) - heap dumps showed that not doing this was a significant problem.