CParser: fix a ClassCastException
After getting a ClassCastException (on init/calibrate.pi), I realized that the
introduction of repOpt in the grammar made the casts I introduced in a rule no
more valid; instead of casting to List[Attribute], now a cast to
List[Opt[Attribute]] is needed; unfortunately, there is no warning for unchecked
casts, nor they fail at runtime - the actual cast is to List, and it succeeds.
Before the casts, there was a pattern match specifying generic type parameters,
which gave a warning, because it looks reasonable and safe and is not.

As a consequence, I updated the types of involved AST nodes.

Not tested - only typechecked.