MultiFeatureContParser: failed attempt at resuming it
This time I could make more sense of some type errors. It became clear that some
errors are deeply rooted into this specific formalization of continuation-based
parsing: for instance, when I resume an old continuation, I need to switch the
result type from ParseResult[T] to ParseResult[OldT], i.e. to the type of the
stored continuations, which is incorrect.