Disable DeMorgan patterns
CNFBenchmark does not complete in a short time with them used. I'm not entirely
sure why, but I believe that if caching worked better, then the extra work would
be done just once per formula, and it would thus become bearable.
Even after using SoftReferences (which are cleared only when short on memory),
this does not work. However, the caching algorithm is known not to guarantee
canonicalization.