FeatureExprLib: cache hashcodes more selectively
Caching hashcodes makes sense only when there's an expensive computation to save, i.e. in And, Or and Not. Save memory in other cases. We might want to remove caching from Not, but it might be not so important.

On And and Or, I tried computing the hash code always at construction time, to avoid a test when accessing it. However, that slowed down testBenchCnf1 from 1.3s-1.5s to ~4s; given that for many objects we skip a lookup in the canonicalization hashtable, and that in many places we generate O(n) temporaries, that's not so surprising. We might want to reduce the number of temporaries we create, as remarked in a comment.

At some point I moved the clauses member into BinaryLogicConnective, for various reasons, and that seemed to give an advantage. However, if there is an advantage, it is smaller than the experimental noise. Therefore, I'll restore the old code in next commit, so that it is kept in the history.