FeatureExprLib: simplify code again as promised in previous commit.
I restored the code after the change I promised to undo. Quoting:
"At some point I moved the clauses member into BinaryLogicConnective, for various reasons, and that seemed to give an advantage. However, if there is an advantage, it is smaller than the experimental noise. Therefore, I'll restore the old code in next commit, so that it is kept in the history."