Preprocessor: prepare for stringify fix - fix autoconverted Scala sources
Now the code builds and passes the testsuite. I did mostly minimal changes to
the source for now, it's still mostly Java with Scala syntax; and fixing that is
not the main point.

Rationale: I converted sources related to stringification, to be able to generalize what an
macro Argument is.