Partial Preprocessor: remove distinction between HASH and '#' to workaround bug.
This distinction creates problem in e.g. kernel/sysctl_binary.c, but actually
everything including include/net/ip_vs.h; the problem is the occurrence of
the ", ## __VA_ARGS__" GNU-C extension, followed by arguments with alternative
expansions, which therefore start with "#if"...; in the current implementation,
"#if" is lexed again, is not at the start of a line, so its '#' is not
recognized as HASH, and therefore refused.

Technically, this might turn some incorrect programs into correct ones, so this
change should be undone, and a proper fix applied (by producing a token sequence
instead of a string, when ", ## __VA_ARGS__" is found, in
MacroTokenSource.paste() and concat()).