PreprocessorTestCase: add more testcases to diagnose the problem with arch/x86/kvm/mmu.c
It turns out that the problem is a bug in the kernel source. It defines
 #define FNAME(arg) a##64_##arg
where FNAME(foo) is supposed to produced an identifier, a64_foo; however, I
believe that according to the C99 standard, the correct output is made of two
identifiers, a64 and _foo, without any space in the middle, but it is still
composed of two separate tokens.
Therefore I'm rewriting the source.