fix test (WTF)
I have no idea why there is a class not found error
without this change when serializing and loading
the same class in a test. The problem is not
reproducable when running form the command line or
within IDEA, but only shows up when running the test
with SBT.
If anybody has any idea, please let me know what could
be a couse of this problem.