skip initial hidden tokens when assigning a position
this is slower but more accurate.

before, the position may point to a too broad text fragment
although part of the text fragment is not included in the
corresponding condition of the result. this is a minor issue
and may not be worth the performance overhead of making this
more accurate.