partially undo change a2ef0fde996aa2
this implementation of checking of GnuAsmExpr has side effects. It produces tons of error messages in Linux.
We should never blindly search for ids. Currently we accept to not understand GnuAsmExpr in details, that's why there is the untyped "stuff" variable. If we want to handle this, we need to handle this properly and change the AST node and properly typecheck it.

Added test that catches this issue