eliminate someDefinition and someImpliedDefinition calls in CDeclUse
This is extremely expensive for performance and is computed even when not
needed at all. Used lots of SAT calls to detemine whether a struct
was defined in a specific context (and only exactly in this context,
which likely does not work correctly in many cases)
Rewrote the code to consider the context when
the struct is undefined. New code does not use any SAT call.

Passes test cases, but not sure that it's actually equivalent.