fight random failures on travis
on travis-ci a few of the hundreds of external CPP calls typically fail for no apparent reason. it's unclear whether it's a problem on travis or with the scala process API or something else.

it's now simply trying to preprocess a file up to three times if the output is empty to work around those problems